LayoutTests/platform/gtk/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-popup-element/popup-light-dismiss.tentative-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_false: expected false got undefined
Popup 1 Outside all popups Inside popup 1  Popup 2 Inside popup 2  Popup 3 - button 3 Inside popup 4 Inside popup 3 Popup 3 - button 4 Inside popup 5  Inside popup 6
Bottom of popup6  Popup 6

Harness Error (FAIL), message = Unhandled rejection: assert_false: expected false got undefined
